pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oscal-agent-lab"
version = "0.4.0"
description = "LangGraph-powered multi-agent copilot for OSCAL catalogs, profiles, and SSPs"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["oscal", "nist", "langgraph", "langchain", "agents", "compliance", "security"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "faiss-cpu>=1.8.0",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
oscal-lab = "oscal_agent_lab.cli:main"
oscal-diff = "oscal_agent_lab.diff_cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/oscal-agent-lab"
Repository = "https://github.com/yourusername/oscal-agent-lab"

[tool.hatch.build.targets.wheel]
packages = ["src/oscal_agent_lab"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
